requirements.txt
flask
torch
torchvision
open-clip-torch
pandas
numpy
scikit-learn
Pillow
pytest